requirements.txt
openai-whisper
sentence-transformers
torch
sounddevice
numpy
pyaudio
faster-whisper
